Remove GETH_NETWORK_ID env-var
geth's networkid is exposed as `GETH_NETWORKID` instead of
`GETH_NETWORK_ID`, so the value has not been used anyways.

This seems to work in either way because the chain-id is read from the
chaindata that gets initialized with the provided genesis file during
`geth init` before the actual node is run.