Make migrate config download safer
It turns out that git does not preserve the original modification time
of a file, so using `wget -N` which only downloads a new copy of a file
if the server modification date is newer than the file's modification
date will not work.

So instead switch to `wget -O` which always overwrites the local file
with the server provided file.